+++ M.A.G.U.S.Assistant/ViewModels/CreatureDetailsViewModel.cs
@@ -68,7 +68,7 @@ internal partial class CreatureDetailsViewModel : BaseViewModel
     public string AstralMagicResistance => !Creature.AstralMagicResistance.HasValue ? "-" : Creature.AstralMagicResistance.Value == Int32.MaxValue ? Lng.Elem("Immune") : Creature.AstralMagicResistance.Value.ToString(CultureInfo.InvariantCulture);
     public string MentalMagicResistance => !Creature.MentalMagicResistance.HasValue ? "-" : Creature.MentalMagicResistance == Int32.MaxValue ? Lng.Elem("Immune") : Creature.MentalMagicResistance.Value.ToString(CultureInfo.InvariantCulture);
     public string PoisonResistance => !Creature.PoisonResistance.HasValue ? "-" : Creature.PoisonResistance == Int32.MaxValue ? Lng.Elem("Immune") : Creature.PoisonResistance.Value.ToString(CultureInfo.InvariantCulture);
-    public uint ExperiencePoints => Creature.ExperiencePoints;
+    public ulong ExperiencePoints => Creature.ExperiencePoints;
     public double AttacksPerRound => Creature.AttacksPerRound;
     public string Image => Creature.Images.Length > 1 ? Creature.Images[RandomProvider.GetSecureRandomInt(0, Creature.Images.Length)] : Creature.Images[0];
     public string Sound => Creature.Sounds.Length > 1 ? Creature.Sounds[RandomProvider.GetSecureRandomInt(0, Creature.Sounds.Length)] : Creature.Sounds[0];
