+++ M.A.G.U.S/Classes/Believer/Priest.cs
@@ -118,7 +118,7 @@ public abstract class Priest : Class, IClass, ILikeMagic
         new() { Level = 12, MinExperience = 130001, MaxExperience = 165000 }
     ];
 
-    public override int ExpPerLevelAfter12 => 50000;
+    public override ulong ExpPerLevelAfter12 => 50000;
 
     public override QualificationList Qualifications => BuildQualifications(
     [
