+++ M.A.G.U.S/Classes/Fighter/Warrior.cs
@@ -111,7 +111,7 @@ public class Warrior : Class, IClass, IJustFight
         new() { Level = 12, MinExperience = 80001, MaxExperience = 112000 }
     ];
 
-    public override int ExpPerLevelAfter12 => 31200;
+    public override ulong ExpPerLevelAfter12 => 31200;
 
     public override IRace[] AllowedRaces => [new Human(), new Elf(), new HalfElf(), new Dwarf(), new CourtOrc(), new Amund(), new Jann(), new Khal(), new Wier(), new Feenhar(), new Dahr(), new Dracker(), new Draquon(),
         new ForestGiant(), new FrostGiant(), new MountainGiant(), new SwampGiant(), new Gnome(), new CourtGoblin()];
