+++ M.A.G.U.S/Classes/NonPlayableCharacters/Jad/JahPriest.cs
@@ -99,7 +99,7 @@ public class JahPriest : Class, IClass
         new() { Level = 12, MinExperience = 130001, MaxExperience = 165000 }
     ];
 
-    public override int ExpPerLevelAfter12 => 50000;
+    public override ulong ExpPerLevelAfter12 => 50000;
 
     public override QualificationList Qualifications => BuildQualifications(
     [
