+++ M.A.G.U.S/Classes/Sorcerer/Warlock.cs
@@ -115,7 +115,7 @@ public class Warlock : Class, IClass, ILikeMagic
         new() { Level = 12, MinExperience = 140001, MaxExperience = 190000 }
     ];
 
-    public override int ExpPerLevelAfter12 => 55000;
+    public override ulong ExpPerLevelAfter12 => 55000;
     
     public override QualificationList Qualifications => BuildQualifications(
     [
