+++ M.A.G.U.S/GameSystem/Experience/LevelRequirement.cs
@@ -4,7 +4,7 @@ public class LevelRequirement
 {
     public int Level { get; set; }
     
-    public int MinExperience { get; set; }
+    public ulong MinExperience { get; set; }
 
-    public int MaxExperience { get; set; }
+    public ulong MaxExperience { get; set; }
 }
